+++ src/views/home/account/pay/Pay.tsx
@@ -1,15 +1,18 @@
-import React, { useState } from 'react';
-import { InputWithDeco } from 'src/components/input/InputWithDeco';
-import { useBosPayMutation } from 'src/graphql/mutations/__generated__/bosPay.generated';
-import { ColorButton } from 'src/components/buttons/colorButton/ColorButton';
-import { Separation, SingleLine } from 'src/components/generic/Styled';
 import { toast } from 'react-toastify';
-import { ChannelSelect } from 'src/components/select/specific/ChannelSelect';
 import { getErrorContent } from 'src/utils/error';
+import { ColorButton } from 'src/components/buttons/colorButton/ColorButton';
+import { SingleLine, Separation } from '../../../../components/generic/Styled';
+import { useEffect, useState } from 'react';
+import { useBosPayMutation } from 'src/graphql/mutations/__generated__/bosPay.generated';
+import { InputWithDeco } from 'src/components/input/InputWithDeco';
+import { ChannelSelect } from 'src/components/select/specific/ChannelSelect';
+import { useDecodeRequestLazyQuery } from 'src/graphql/queries/__generated__/decodeRequest.generated';
+import { renderLine } from 'src/components/generic/helpers';
+import { Price } from 'src/components/price/Price';
+import { Camera } from 'react-feather';
 import Modal from 'src/components/modal/ReactModal';
 import dynamic from 'next/dynamic';
 import { LoadingCard } from 'src/components/loading/LoadingCard';
-import { Camera } from 'react-feather';
 
 const QRCodeReader = dynamic(() => import('src/components/qrReader'), {
   ssr: false,
@@ -18,18 +21,23 @@ const QRCodeReader = dynamic(() => import('src/components/qrReader'), {
   },
 });
 
-type PayProps = {
+interface PayProps {
   predefinedRequest?: string;
   payCallback?: () => void;
-};
+}
 
-export const Pay = ({ predefinedRequest, payCallback }: PayProps) => {
+export const Pay: React.FC<PayProps> = ({ predefinedRequest, payCallback }) => {
   const [modalOpen, setModalOpen] = useState<boolean>(false);
 
   const [request, setRequest] = useState<string>(predefinedRequest || '');
   const [peers, setPeers] = useState<string[]>([]);
-  const [fee, setFee] = useState<number>(1);
+  const [fee, setFee] = useState<number>(10);
   const [paths, setPaths] = useState<number>(1);
+
+  const [decode, { data, loading: decodeLoading }] = useDecodeRequestLazyQuery({
+    fetchPolicy: 'network-only',
+    onError: () => toast.error('Error decoding invoice'),
+  });
 
   const [pay, { loading }] = useBosPayMutation({
     onCompleted: () => {
@@ -39,6 +47,12 @@ export const Pay = ({ predefinedRequest, payCallback }: PayProps) => {
     },
     onError: error => toast.error(getErrorContent(error)),
   });
+
+  useEffect(() => {
+    if (predefinedRequest) {
+      decode({ variables: { request: predefinedRequest } });
+    }
+  }, []);
 
   const handleEnter = () => {
     if (loading || !request) return;
@@ -52,6 +66,27 @@ export const Pay = ({ predefinedRequest, payCallback }: PayProps) => {
     });
   };
 
+  const renderDecoded = () => {
+    if (decodeLoading) {
+      return <>Decoding invoice...</>;
+    }
+
+    if (!data?.decodeRequest) return null;
+
+    const { description, tokens, destination_node } = data.decodeRequest;
+
+    const { alias } = destination_node.node;
+
+    return (
+      <>
+        {renderLine('Description', description)}
+        {renderLine('Value', <Price amount={tokens} />)}
+        {renderLine('Destination', alias)}
+        <Separation />
+      </>
+    );
+  };
+
   return (
     <>
       {!predefinedRequest && (
@@ -63,7 +98,11 @@ export const Pay = ({ predefinedRequest, payCallback }: PayProps) => {
               placeholder={'Invoice'}
               inputCallback={value => setRequest(value)}
               onEnter={() => handleEnter()}
-              inputMaxWidth={'440px'}
+              inputMaxWidth={'240px'}
+              blurCallback={() => {
+                if (!request) return;
+                decode({ variables: { request } });
+              }}
             />
             <ColorButton
               withMargin={'0 0 0 8px'}
@@ -75,12 +114,14 @@ export const Pay = ({ predefinedRequest, payCallback }: PayProps) => {
           <Separation />
         </>
       )}
+      {renderDecoded()}
       <InputWithDeco
         title={'Max Fee'}
         value={fee}
         amount={fee}
         placeholder={'sats'}
         inputType={'number'}
+        inputMaxWidth={'300px'}
         inputCallback={value => setFee(Math.max(1, Number(value)))}
         onEnter={() => handleEnter()}
       />
@@ -89,6 +130,7 @@ export const Pay = ({ predefinedRequest, payCallback }: PayProps) => {
         value={paths}
         placeholder={'paths'}
         inputType={'number'}
+        inputMaxWidth={'300px'}
         inputCallback={value => setPaths(Math.max(1, Number(value)))}
         onEnter={() => handleEnter()}
       />
@@ -96,6 +138,7 @@ export const Pay = ({ predefinedRequest, payCallback }: PayProps) => {
       <ChannelSelect
         title={'Out Channels'}
         isMulti={true}
+        maxWidth={'300px'}
         callback={p => setPeers(p.map(peer => peer.partner_public_key))}
       />
       <Separation />